+++ App/Repositories/Admin/PageRepository.php
@@ -114,7 +114,9 @@ class PageRepository
             $target_dir = Config::F_FEATURED_IMAGE_ROOT;
             $target_file = $target_dir . $post[0]['featured_image'];
             try {
-                unlink($target_file);
+                if (file_exists($target_file)) {
+                    unlink($target_file);
+                }
             } catch (Exception $e) {
 
             }
